feat: add BaseUrl feature
add the ability to set a BaseUrl on the web client, which is sent to the backend as a header named x-base-url.
the backend then sends requests to the corresponding BaseUrl based on this header.
when the client doesn't set an API key or BaseUrl (i.e. when accessed via code), the server will use the pre-set BaseUrl in the environment variable to send requests to the corresponding address.

Signed-off-by: Homing So <homingso@foxmail.com>