[+] Fix Export all data functionality to save state as a JSON file instead of overwriting it as a .txt file.
The previous code provided a hook called useSyncStore that included an export function for exporting the state. However, the export functionality was overwriting the state as a .txt file (Tested in iPhone Browser : Safari), which was not desirable. To address this issue, the code was modified to export the state as a JSON file instead.