SyncService [webdav]
[+] Update `check` method to use `PROPFIND` instead of `MKCOL`

In the `webdav.ts` module, updated the `check` method to use the `PROPFIND` method instead of `MKCOL` to check if the file exists. This change was made to handle cases where the server does not allow the `MKCOL` method. The `PROPFIND` method is used to retrieve the properties of the file, and if the response status is either 200 (OK), 207 (Multi-Status), or 404 (Not Found), it considers the file to exist.